requirements.txt
flask==3.0.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
playwright>=1.40.0
python-dotenv>=1.0.0
